# swagger/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "title": "SeqUI Server",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/seqapi/v1/async_search/list": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "seqapi_v1"
                ],
                "operationId": "seqapi_v1_get_async_searches_list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.GetAsyncSearchesListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.GetAsyncSearchesListResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
...
        "/seqapi/v1/histogram": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "seqapi_v1"
                ],
                "operationId": "seqapi_v1_getHistogram",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.GetHistogramRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.GetHistogramResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
        "/seqapi/v1/limits": {
            "get": {
                "tags": [
                    "seqapi_v1"
                ],
                "operationId": "seqapi_v1_getLimits",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.GetLimitsResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
        "/seqapi/v1/logs_lifespan": {
            "get": {
                "tags": [
                    "seqapi_v1"
                ],
                "operationId": "seqapi_v1_get_logs_lifespan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.GetLogsLifespanResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "UnexpectedError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "http.AsyncSearchRequestHistogram": {
            "type": "object",
            "properties": {
                "interval": {
                    "type": "string"
                }
            }
        },
...
        "seqapi.v1.AggregationFunc": {
            "type": "string",
            "enum": [
                "count",
                "sum",
                "min",
                "max",
                "avg",
                "quantile",
                "unique"
            ],
            "x-enum-varnames": [
                "afCount",
                "afSum",
                "afMin",
                "afMax",
                "afAvg",
                "afQuantile",
                "afUnique"
            ]
        },
...
        "seqapi.v1.AggregationTsQuery": {
            "type": "object",
            "properties": {
                "agg_func": {
                    "default": "count",
                    "allOf": [
                        {
                            "$ref": "#/definitions/seqapi.v1.AggregationFunc"
                        }
                    ]
                },
                "field": {
                    "type": "string"
                },
                "group_by": {
                    "type": "string"
                },
                "interval": {
                    "type": "string",
                    "format": "duration",
                    "example": "1m"
                },
                "quantiles": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
...
        "seqapi.v1.AsyncSearchStatus": {
            "type": "string",
            "enum": [
                "in_progress",
                "done",
                "canceled",
                "error"
            ],
            "x-enum-varnames": [
                "AsyncSearchStatusInProgress",
                "AsyncSearchStatusDone",
                "AsyncSearchStatusCanceled",
                "AsyncSearchStatusError"
            ]
        },
        "seqapi.v1.AsyncSearchesListItem": {
            "type": "object",
            "properties": {
                "canceled_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "disk_usage": {
                    "type": "string",
                    "format": "int64"
                },
                "error": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "owner_name": {
                    "type": "string"
                },
                "progress": {
                    "type": "number"
                },
                "request": {
                    "$ref": "#/definitions/seqapi.v1.StartAsyncSearchRequest"
                },
                "search_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "started_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "status": {
                    "$ref": "#/definitions/seqapi.v1.AsyncSearchStatus"
                }
            }
        },
        "seqapi.v1.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "default": "ERROR_CODE_NO",
                    "allOf": [
                        {
                            "$ref": "#/definitions/seqapi.v1.ErrorCode"
                        }
                    ]
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "seqapi.v1.ErrorCode": {
            "type": "string",
            "enum": [
                "ERROR_CODE_NO",
                "ERROR_CODE_PARTIAL_RESPONSE",
                "ERROR_CODE_QUERY_TOO_HEAVY",
                "ERROR_CODE_TOO_MANY_FRACTIONS_HIT"
            ],
            "x-enum-varnames": [
                "aecNo",
                "aecPartialResponse",
                "aecQueryTooHeavy",
                "aecTooManyFractionsHit"
            ]
        },
...
        "seqapi.v1.GetAsyncSearchesListRequest": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "format": "int32"
                },
                "offset": {
                    "type": "integer",
                    "format": "int32"
                },
                "owner_name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/seqapi.v1.AsyncSearchStatus"
                }
            }
        },
        "seqapi.v1.GetAsyncSearchesListResponse": {
            "type": "object",
            "properties": {
                "searches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.AsyncSearchesListItem"
                    }
                }
            }
        },
...
        "seqapi.v1.GetHistogramRequest": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string",
                    "format": "date-time"
                },
                "interval": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "to": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "seqapi.v1.GetHistogramResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/seqapi.v1.Error"
                },
                "histogram": {
                    "$ref": "#/definitions/seqapi.v1.Histogram"
                },
                "partialResponse": {
                    "type": "boolean"
                }
            }
        },
        "seqapi.v1.GetLimitsResponse": {
            "type": "object",
            "properties": {
                "maxAggregationsPerRequest": {
                    "type": "integer",
                    "format": "int32"
                },
                "maxExportLimit": {
                    "type": "integer",
                    "format": "int32"
                },
                "maxParallelExportRequests": {
                    "type": "integer",
                    "format": "int32"
                },
                "maxSearchLimit": {
                    "type": "integer",
                    "format": "int32"
                },
                "seqCliMaxSearchLimit": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
        "seqapi.v1.GetLogsLifespanResponse": {
            "type": "object",
            "properties": {
                "lifespan": {
                    "type": "integer"
                }
            }
        },
        "seqapi.v1.Histogram": {
            "type": "object",
            "properties": {
                "buckets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.HistogramBucket"
                    }
                }
            }
        },
        "seqapi.v1.HistogramBucket": {
            "type": "object",
            "properties": {
                "docCount": {
                    "type": "string",
                    "format": "uint64"
                },
                "key": {
                    "type": "string",
                    "format": "uint64"
                }
            }
        },
...
        "seqapi.v1.StartAsyncSearchRequest": {
            "type": "object",
            "properties": {
                "aggregations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.AggregationTsQuery"
                    }
                },
                "from": {
                    "type": "string",
                    "format": "date-time"
                },
                "histogram": {
                    "$ref": "#/definitions/http.AsyncSearchRequestHistogram"
                },
                "meta": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "retention": {
                    "type": "string",
                    "format": "duration",
                    "example": "1h"
                },
                "size": {
                    "type": "integer"
                },
                "to": {
                    "type": "string",
                    "format": "date-time"
                },
                "with_docs": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearer": {
            "description": "Authentication token, prefixed by Bearer: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}